Move os.h from source/Irrlicht to include/
